Refactor item_randomizer and item_shuffler
